NSPX Comit 8
1. Actor animation fix (idle without weapon)  ActorAnimation.cpp
2. Vsynh fix in HW.cpp

+++ xr_3da/xrGame/MathUtils.h
@@ -235,14 +235,14 @@ IC	void twoq_2w(const Fquaternion& q1,const Fquaternion& q2,float dt,Fvector& w)
 	Fvector v1,v2;
 	v1.set(q1.x,q1.y,q1.z);
 	v2.set(q2.x,q2.y,q2.z);
-	float cosinus=q1.w*q2.w+v1.dotproduct(v2);//q1.w*q2.w+ q1.v.dotproduct(q2.v)
+	double cosinus=q1.w*q2.w+v1.dotproduct(v2);//q1.w*q2.w+ q1.v.dotproduct(q2.v)
 	w.crossproduct(v1,v2);
 	//								  //the signum must be inverted ?
 	v1.mul(q2.w);
 	v2.mul(q1.w);
 	w.sub(v2);
 	w.add(v1);
-	float sinus_2=1.f-cosinus*cosinus,k=2.f/dt;
+	double sinus_2=1.f-cosinus*cosinus,k=2.f/dt;
 	if(sinus_2>EPS)	k*=acos(cosinus)/_sqrt(sinus_2);
 	w.mul(k);
 }
